Trait interval::ops::Width

pub trait Width: Ord + Clone {
    type Output: Unsigned + Integer + Clone;
    fn max_value() -> Self;

    fn min_value() -> Self;

    fn width(lower: &Self, upper: &Self) -> Self::Output;
}

Limit of a bound for which the distance between min_value() and max_value() can be represented in the type Output.

Associated Types

type Output: Unsigned + Integer + Clone

Loading content...
Required methods

fn max_value() -> Self

fn min_value() -> Self

fn width(lower: &Self, upper: &Self) -> Self::Output

The result might be infinite depending on the underlying type (think about floating types).

Loading content...